+++ dev/breeze/doc/images/output_build-docs.txt
@@ -1,1 +1,1 @@
-8c3469cb20447510da111e0e59c97ae2
+7b9620b5ebb99d23be778d8632da8150
