+++ dev/breeze/doc/images/output_release-management_add-back-references.txt
@@ -1,1 +1,1 @@
-28ca3febf77042587cb889cc72053872
+12a24f48d706bfdc9893951274e5574f
