+++ dev/breeze/doc/images/output_release-management_publish-docs.txt
@@ -1,1 +1,1 @@
-34ecaf7e65f8beecd2086b093b4f4e1e
+160c1fa48160b72ece23b6793347a9b9
