+++ dev/breeze/doc/images/output_sbom_generate-providers-requirements.txt
@@ -1,1 +1,1 @@
-2b7b93625b6e3adc20712223eb230494
+f1332e8ce3cc88d0ea9ece1e6ddeec76
