+++ providers/git/src/airflow/providers/git/hooks/__init__.py
@@ -14,34 +14,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
----
-package-name: apache-airflow-providers-git
-name: GIT
-description: |
-    `Distributed version control system (GIT) <https://git-scm.com/>`__
-
-state: not-ready
-source-date-epoch: 1726861127
-# note that those versions are maintained by release manager - do not update them manually
-versions:
-  - 1.0.0
-
-integrations:
-  - integration-name: GIT (Git)
-    external-doc-url: https://git-scm.com/
-    tags: [software]
-
-hooks:
-  - integration-name: GIT
-    python-modules:
-      - airflow.providers.git.hooks.git
-bundles:
-  - integration-name: GIT
-    python-modules:
-      - airflow.providers.git.bundles.git
-
-
-connection-types:
-  - hook-class-name: airflow.providers.git.hooks.git.GitHook
-    connection-type: git
