+++ pyproject.toml
@@ -102,7 +102,7 @@ dynamic = ["version", "optional-dependencies", "dependencies"]
 # apache.hdfs, apache.hive, apache.iceberg, apache.impala, apache.kafka, apache.kylin, apache.livy,
 # apache.pig, apache.pinot, apache.spark, apprise, arangodb, asana, atlassian.jira, celery, cloudant,
 # cncf.kubernetes, cohere, common.compat, common.io, common.messaging, common.sql, databricks,
-# datadog, dbt.cloud, dingding, discord, docker, edge, elasticsearch, exasol, fab, facebook, ftp,
+# datadog, dbt.cloud, dingding, discord, docker, edge, elasticsearch, exasol, fab, facebook, ftp, git,
 # github, google, grpc, hashicorp, http, imap, influxdb, jdbc, jenkins, microsoft.azure,
 # microsoft.mssql, microsoft.psrp, microsoft.winrm, mongo, mysql, neo4j, odbc, openai, openfaas,
 # openlineage, opensearch, opsgenie, oracle, pagerduty, papermill, pgvector, pinecone, postgres,
@@ -791,6 +791,101 @@ dev = [
     "apache-airflow[all-core]",
     "apache-airflow-task-sdk",
     "apache-airflow-devel-common",
+    "apache-airflow-providers-airbyte",
+    "apache-airflow-providers-alibaba",
+    "apache-airflow-providers-amazon",
+    "apache-airflow-providers-apache-beam",
+    "apache-airflow-providers-apache-cassandra",
+    "apache-airflow-providers-apache-drill",
+    "apache-airflow-providers-apache-druid",
+    "apache-airflow-providers-apache-flink",
+    "apache-airflow-providers-apache-hdfs",
+    "apache-airflow-providers-apache-hive",
+    "apache-airflow-providers-apache-iceberg",
+    "apache-airflow-providers-apache-impala",
+    "apache-airflow-providers-apache-kafka",
+    "apache-airflow-providers-apache-kylin",
+    "apache-airflow-providers-apache-livy",
+    "apache-airflow-providers-apache-pig",
+    "apache-airflow-providers-apache-pinot",
+    "apache-airflow-providers-apache-spark",
+    "apache-airflow-providers-apprise",
+    "apache-airflow-providers-arangodb",
+    "apache-airflow-providers-asana",
+    "apache-airflow-providers-atlassian-jira",
+    "apache-airflow-providers-celery",
+    "apache-airflow-providers-cloudant",
+    "apache-airflow-providers-cncf-kubernetes",
+    "apache-airflow-providers-cohere",
+    "apache-airflow-providers-common-compat",
+    "apache-airflow-providers-common-io",
+    "apache-airflow-providers-common-messaging",
+    "apache-airflow-providers-common-sql",
+    "apache-airflow-providers-databricks",
+    "apache-airflow-providers-datadog",
+    "apache-airflow-providers-dbt-cloud",
+    "apache-airflow-providers-dingding",
+    "apache-airflow-providers-discord",
+    "apache-airflow-providers-docker",
+    "apache-airflow-providers-edge",
+    "apache-airflow-providers-elasticsearch",
+    "apache-airflow-providers-exasol",
+    "apache-airflow-providers-fab",
+    "apache-airflow-providers-facebook",
+    "apache-airflow-providers-ftp",
+    "apache-airflow-providers-git",
+    "apache-airflow-providers-github",
+    "apache-airflow-providers-google",
+    "apache-airflow-providers-grpc",
+    "apache-airflow-providers-hashicorp",
+    "apache-airflow-providers-http",
+    "apache-airflow-providers-imap",
+    "apache-airflow-providers-influxdb",
+    "apache-airflow-providers-jdbc",
+    "apache-airflow-providers-jenkins",
+    "apache-airflow-providers-microsoft-azure",
+    "apache-airflow-providers-microsoft-mssql",
+    "apache-airflow-providers-microsoft-psrp",
+    "apache-airflow-providers-microsoft-winrm",
+    "apache-airflow-providers-mongo",
+    "apache-airflow-providers-mysql",
+    "apache-airflow-providers-neo4j",
+    "apache-airflow-providers-odbc",
+    "apache-airflow-providers-openai",
+    "apache-airflow-providers-openfaas",
+    "apache-airflow-providers-openlineage",
+    "apache-airflow-providers-opensearch",
+    "apache-airflow-providers-opsgenie",
+    "apache-airflow-providers-oracle",
+    "apache-airflow-providers-pagerduty",
+    "apache-airflow-providers-papermill",
+    "apache-airflow-providers-pgvector",
+    "apache-airflow-providers-pinecone",
+    "apache-airflow-providers-postgres",
+    "apache-airflow-providers-presto",
+    "apache-airflow-providers-qdrant",
+    "apache-airflow-providers-redis",
+    "apache-airflow-providers-salesforce",
+    "apache-airflow-providers-samba",
+    "apache-airflow-providers-segment",
+    "apache-airflow-providers-sendgrid",
+    "apache-airflow-providers-sftp",
+    "apache-airflow-providers-singularity",
+    "apache-airflow-providers-slack",
+    "apache-airflow-providers-smtp",
+    "apache-airflow-providers-snowflake",
+    "apache-airflow-providers-sqlite",
+    "apache-airflow-providers-ssh",
+    "apache-airflow-providers-standard",
+    "apache-airflow-providers-tableau",
+    "apache-airflow-providers-telegram",
+    "apache-airflow-providers-teradata",
+    "apache-airflow-providers-trino",
+    "apache-airflow-providers-vertica",
+    "apache-airflow-providers-weaviate",
+    "apache-airflow-providers-yandex",
+    "apache-airflow-providers-ydb",
+    "apache-airflow-providers-zendesk",
 ]
 
 [tool.uv]
@@ -845,6 +940,7 @@ apache-airflow-providers-exasol = { workspace = true }
 apache-airflow-providers-fab = { workspace = true }
 apache-airflow-providers-facebook = { workspace = true }
 apache-airflow-providers-ftp = { workspace = true }
+apache-airflow-providers-git = { workspace = true }
 apache-airflow-providers-github = { workspace = true }
 apache-airflow-providers-google = { workspace = true }
 apache-airflow-providers-grpc = { workspace = true }
@@ -947,6 +1043,7 @@ members = [
     "providers/fab",
     "providers/facebook",
     "providers/ftp",
+    "providers/git",
     "providers/github",
     "providers/google",
     "providers/grpc",
