+++ scripts/ci/docker-compose/remove-sources.yml
@@ -68,6 +68,7 @@ services:
       - ../../../empty:/opt/airflow/providers/fab/src
       - ../../../empty:/opt/airflow/providers/facebook/src
       - ../../../empty:/opt/airflow/providers/ftp/src
+      - ../../../empty:/opt/airflow/providers/git/src
       - ../../../empty:/opt/airflow/providers/github/src
       - ../../../empty:/opt/airflow/providers/google/src
       - ../../../empty:/opt/airflow/providers/grpc/src
